Moving enums from Esfa.Recruit.Vacancies.Client.Domain.Enums to Esfa.Recruit.Vacancies.Client.Domain.Entities
