+++ src/Employer/Employer.Web/Orchestrators/Part2/VacancyDescriptionOrchestrator.cs
@@ -1,6 +1,4 @@
 ﻿using Esfa.Recruit.Employer.Web.ViewModels;
-using Esfa.Recruit.Vacancies.Client.Domain;
-using Esfa.Recruit.Vacancies.Client.Domain.Enums;
 using Esfa.Recruit.Vacancies.Client.Domain.Exceptions;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Client;
 using System;
