+++ src/Employer/UnitTests/Core/Entities/VacancyExtensionsTest.cs
@@ -1,5 +1,4 @@
 ﻿using Esfa.Recruit.Vacancies.Client.Domain.Entities;
-using Esfa.Recruit.Vacancies.Client.Domain.Enums;
 using FluentAssertions;
 using Xunit;
