+++ src/Shared/Recruit.Vacancies.Client/Domain/Entities/DurationUnit.cs
@@ -1,4 +1,4 @@
-﻿namespace Esfa.Recruit.Vacancies.Client.Domain.Enums
+﻿namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
 {
     public enum DurationUnit
     {
