+++ src/Shared/Recruit.Vacancies.Client/Domain/Entities/Programme.cs
@@ -1,6 +1,4 @@
-﻿using Esfa.Recruit.Vacancies.Client.Domain.Enums;
-
-namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
+﻿namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
 {
     public class Programme
     {
