+++ src/Shared/Recruit.Vacancies.Client/Domain/Entities/Qualification.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Esfa.Recruit.Vacancies.Client.Domain.Enums;
-
-namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
+﻿namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
 {
     public class Qualification
     {
