+++ src/Shared/Recruit.Vacancies.Client/Domain/Projections/VacancySummary.cs
@@ -1,5 +1,5 @@
-﻿using Esfa.Recruit.Vacancies.Client.Domain.Enums;
-using System;
+﻿using System;
+using Esfa.Recruit.Vacancies.Client.Domain.Entities;
 
 namespace Esfa.Recruit.Vacancies.Client.Domain.Projections
 {
