+++ src/Shared/UnitTests/Application/VacancyValidation/CrossField/FixedWageAmountValidation.cs
@@ -1,7 +1,6 @@
 using System;
 using Esfa.Recruit.Vacancies.Client.Application.Validation;
 using Esfa.Recruit.Vacancies.Client.Domain.Entities;
-using Esfa.Recruit.Vacancies.Client.Domain.Enums;
 using FluentAssertions;
 using Moq;
 using Xunit;
