+++ src/Shared/UnitTests/Application/VacancyValidation/SingleField/QualificationsTests.cs
@@ -4,7 +4,6 @@ using System.Text;
 using Esfa.Recruit.Vacancies.Client.Application.Configuration;
 using Esfa.Recruit.Vacancies.Client.Application.Validation;
 using Esfa.Recruit.Vacancies.Client.Domain.Entities;
-using Esfa.Recruit.Vacancies.Client.Domain.Enums;
 using FluentAssertions;
 using Xunit;
