+++ src/Shared/UnitTests/Application/VacancyValidation/SingleField/TrainingValidationTests.cs
@@ -1,6 +1,5 @@
 using Esfa.Recruit.Vacancies.Client.Application.Validation;
 using Esfa.Recruit.Vacancies.Client.Domain.Entities;
-using Esfa.Recruit.Vacancies.Client.Domain.Enums;
 using FluentAssertions;
 using Xunit;
